SVN: Test case 119 labels_branch-branch.
Test what happens when a label is branched with edits and deletes.  The default behaviour
is to keep the initial label as a label and the resulting branch is sparse.  If this is 
not desired then the tag should be included in the branch mapping (include.map).

For example:

[exclude.map]
^tags/.*

[include.map]
^tags/full_branch/.* 

Copied from Perforce
 Change: 11047