Path map translator for CVS and SVN paths.
(undoc)
To use create a path.map file with regex and group match.  Only the first matching entry is used.  The regex and group match are seperated by ', ' (or in regex terms ',\s+').  Lines starting with '#' are ignored.

For example, 'trunk' is renamed to 'main', but other entries are left as-is.

# path.map
trunk/(.*), //import/main/{1}
(.*), //import/{1}

Note: if no file is found the default 'depot' and 'subPath' options are used to generate the map, preserving the original behaviour.

CVS paths will always stat with the 'branch' name.  'main' for 1.1 and the symbol for other branches.

# path.map
main/projA/(.*), //import/projA/MAIN/{1}
release_(.*)/projA/(.*), //import/projA/REL{1}/{2}
(.*)/projA/(.*), //import/projA/TAG-{1}/{2}
(.*), //import/unexpected/{1}

Node: adding a catch all is a good idea.

Copied from Perforce
 Change: 11071