New non-translate mode for high-ascii files.
If the P4_TRANSLATE mode is disabled then text files with high-ascii characters are given the new type RAW-TEXT. Raw types are not translated and the content is used as-is. (exception UTF16/32)

The default translation configuration is enabled:

   com.p4convert.p4.translate=true

The com.p4convert.p4.translate=false mode is intended for use with non-unicode Perforce servers in a Windows only client environment.

Copied from Perforce
 Change: 12222